requirements.txt
flask==3.1.3
flask-cors==6.0.2
openai==2.21.0
python-dotenv==1.2.1
pydantic==2.12.5
requests==2.32.5